extract the pinned cell and related components
