+++ packages/core/src/components/index.js
@@ -1,4 +1,6 @@
-import Notebook from "./notebook";
+// @flow
+// TODO: Move this outta here
+import Notebook from "../providers/notebook";
 
 const _nextgen = require("./ng");
